feat: Support four parameters (u, v, uOpposite and vOpposite) (#66)
This PR makes significant changes to the API and underlying code to accept for parameters instead of the two that are usually used for a coons patch. This greatly increases the flexibility of the interpolation as now points on opposing boundaries do not need to be opposites.

It includes updated tests and documentation.

BREAKING CHANGE: This will require a major version bump as there is a change in the types of the arguments.